+++ packages/editor/editor-wikimarkup-transformer/src/parser/macro/panel.ts
@@ -4,6 +4,7 @@ import { normalizePMNodes } from '../utils/normalize';
 import { title } from '../utils/title';
 
 import { getPanelType } from './panel-type';
+import { TokenErrCallback } from '../tokenize';
 
 const allowedNodeType = ['paragraph', 'heading', 'orderedList', 'bulletList'];
 
@@ -11,6 +12,7 @@ export function panelMacro(
   attrs: { [key: string]: string },
   rawContent: string,
   schema: Schema,
+  tokenErrCallback?: TokenErrCallback,
 ): PMNode[] {
   const output: PMNode[] = [];
 
@@ -23,7 +25,7 @@ export function panelMacro(
     output.push(title(attrs.title, schema));
   }
 
-  const content = parseString(rawContent, schema);
+  const content = parseString(rawContent, schema, [], tokenErrCallback);
   const normalizedContent = normalizePMNodes(content, schema);
   let contentBuffer: PMNode[] = [];