+++ packages/search/global-search/src/__tests__/unit/CrossProductSearchClient.test.ts
@@ -1,5 +1,6 @@
 import CrossProductSearchClient, {
   CrossProductSearchResponse,
+  SearchSession,
 } from '../../api/CrossProductSearchClient';
 import { Scope, ConfluenceItem, PersonItem } from '../../api/types';
 import 'whatwg-fetch';
@@ -27,6 +28,11 @@ function apiWillReturn(state: CrossProductSearchResponse) {
 
   fetchMock.mock('localhost/quicksearch/v1', state, opts);
 }
+
+const searchSession: SearchSession = {
+  referrerId: 'referal-id',
+  sessionId: 'test_uuid',
+};
 
 describe('CrossProductSearchClient', () => {
   let searchClient: CrossProductSearchClient;
@@ -61,7 +67,7 @@ describe('CrossProductSearchClient', () => {
         ],
       });
 
-      const result = await searchClient.search('query', 'test_uuid', [
+      const result = await searchClient.search('query', searchSession, [
         Scope.ConfluencePageBlog,
       ]);
       expect(result.results.get(Scope.ConfluencePageBlog)).toHaveLength(1);
@@ -106,7 +112,7 @@ describe('CrossProductSearchClient', () => {
         ],
       });
 
-      const result = await searchClient.search('query', 'test_uuid', [
+      const result = await searchClient.search('query', searchSession, [
         Scope.ConfluenceSpace,
       ]);
       expect(result.results.get(Scope.ConfluenceSpace)).toHaveLength(1);
@@ -149,7 +155,7 @@ describe('CrossProductSearchClient', () => {
         ],
       });
 
-      const result = await searchClient.search('query', 'test_uuid', [
+      const result = await searchClient.search('query', searchSession, [
         Scope.JiraIssue,
       ]);
       expect(result.results.get(Scope.JiraIssue)).toHaveLength(1);
@@ -189,7 +195,7 @@ describe('CrossProductSearchClient', () => {
 
       const result = await searchClient.search(
         'query',
-        'test_uuid',
+        searchSession,
         jiraScopes,
       );
       expect(result.results.get(Scope.JiraIssue)).toHaveLength(0);
@@ -263,7 +269,7 @@ describe('CrossProductSearchClient', () => {
       ],
     });
 
-    const result = await searchClient.search('query', 'test_uuid', [
+    const result = await searchClient.search('query', searchSession, [
       Scope.ConfluencePageBlog,
       Scope.ConfluenceSpace,
     ]);
@@ -277,7 +283,7 @@ describe('CrossProductSearchClient', () => {
       scopes: [],
     });
     // @ts-ignore
-    const result = await searchClient.search('query', 'test_uuid', [
+    const result = await searchClient.search('query', searchSession, [
       Scope.ConfluencePageBlog,
       Scope.JiraIssue,
     ]);